requirements.txt
dj-database-url==0.2.2
Django==1.7.2
django-filter==0.9.2
djangorestframework==3.0.3
Markdown==2.5.2
requests==1.2.3
